// src/lib/api/generated/models/v1OrderType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum v1OrderType {
    ORDER_TYPE_UNSPECIFIED = 'ORDER_TYPE_UNSPECIFIED',
    ORDER_TYPE_MARKET = 'ORDER_TYPE_MARKET',
    ORDER_TYPE_LIMIT = 'ORDER_TYPE_LIMIT',
    ORDER_TYPE_STOP = 'ORDER_TYPE_STOP',
    ORDER_TYPE_STOP_LIMIT = 'ORDER_TYPE_STOP_LIMIT',
}